pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dt-sandboxes"
version = "0.1.0"
description = "Deepthought S.T.E.A.M. Crucibles: Kaplay.js, Minetest, TurboWarp wrappers."
requires-python = ">=3.12,<3.13"
authors = [{ name = "Epistemon" }]
dependencies = [
    "dt-contracts",
    "dt-testing",
    "pydantic>=2.10.0",
]

[tool.uv]
managed = true

[tool.uv.sources]
dt-contracts = { workspace = true }
dt-testing = { workspace = true }

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
]

[tool.ruff]
line-length = 110
target-version = "py314"
exclude = ["vendors"]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "E501", "ISC001", "D203", "D213", "TRY003", "TRY300", "TRY301",
    "A002", "PLR0913", "S101", "BLE001", "PLW0603", "SLF001"
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "PLC0415", "ANN", "ARG001", "PLR2004", "PERF401"]
"src/dt_sandboxes/schemas.py" = ["TC002"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "architecture: architectural compliance tests",
]
filterwarnings = ["error"]

[tool.coverage.report]
fail_under = 90
show_missing = true
